refactor(images): refactor image crafters, will add location
