+++ jina/executors/crafters/image/resize.py
@@ -6,17 +6,19 @@ from typing import Union, Tuple, Dict
 
 import numpy as np
 
-from . import ImageChunkCrafter
+from .. import BaseCrafter
+from .helper import _load_image, _restore_channel_axis, _resize_short
 
 
-class ImageResizer(ImageChunkCrafter):
+class ImageResizer(BaseCrafter):
     """
     :class:`ImageResizer` resize the image to the given size.
     """
 
     def __init__(self,
-                 target_size: Union[Tuple[int], int],
-                 how='BILINEAR',
+                 target_size: Union[Tuple[int, int], int],
+                 how: str = 'BILINEAR',
+                 channel_axis: int = -1,
                  *args, **kwargs):
         """
 
@@ -32,6 +34,7 @@ class ImageResizer(ImageChunkCrafter):
         else:
             raise ValueError(f'output_dim {target_size} should be an integer')
         self.how = how
+        self.channel_axis = channel_axis
 
     def craft(self, blob: 'np.ndarray', *args, **kwargs) -> Dict:
         """
@@ -40,7 +43,7 @@ class ImageResizer(ImageChunkCrafter):
         :param blob: the ndarray of the image
         :return: a chunk dict with the cropped image
         """
-        raw_img = self.load_image(blob)
-        _img = self._resize_short(raw_img, self.output_dim, self.how)
-        img = self.restore_channel_axis(np.asarray(_img))
+        raw_img = _load_image(blob, self.channel_axis)
+        _img = _resize_short(raw_img, self.output_dim, self.how)
+        img = _restore_channel_axis(np.asarray(_img), self.channel_axis)
         return dict(offset=0, weight=1., blob=img.astype('float32'))
